+++ elections/src/view.rs
@@ -34,9 +34,7 @@ impl Contract {
 
     /// Returns the policy if user has accepted it otherwise returns None
     pub fn accepted_policy(&self, user: AccountId) -> Option<String> {
-        self.accepted_policy
-            .get(&user)
-            .map(hex::encode)
+        self.accepted_policy.get(&user).map(hex::encode)
     }
 
     /// Returns all the users votes for all the proposals. If user has not voted yet a vector with None values will be returned.
@@ -90,7 +88,7 @@ impl Contract {
         indexed_results.sort_by_key(|&(_, value)| std::cmp::Reverse(value));
 
         let mut winners = Vec::new();
-        for (idx, votes) in indexed_results[0..=proposal.seats as usize].iter() {
+        for (idx, votes) in indexed_results[0..proposal.seats as usize].iter() {
             if votes >= &proposal.min_candidate_support {
                 let candidate = proposal.candidates.get(*idx).unwrap();
                 winners.push(candidate.clone());